REFACTOR: logger and overrides, WIP
